- update deploy script

+++ deploy.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-rm -R ~/.config/conky/hybrid-compact/
-echo 'hybrid-compact uninstalled'
+if [ -d ~/.config/conky/hybrid-compact/ ]
+then
+    rm -R ~/.config/conky/hybrid-compact/
+    echo 'hybrid-compact uninstalled'
+fi
 
 rsync -IrW --stats --exclude={'.git','deploy.sh','readme.md','screenshots'} $(pwd) ~/.config/conky/
 echo 'hybrid-compact installed'
